requirements.txt
fastapi
uvicorn
pydantic
asyncpg
numpy
openai
# scikit-learn
# nltk
# transformers
# torch
PyPDF2
python-docx
python-pptx